+++ src/main/java/com/example/medicare_call/controller/action/CareCallSettingController.java
@@ -1,7 +1,7 @@
 package com.example.medicare_call.controller.action;
 
 import com.example.medicare_call.dto.CareCallSettingRequest;
-import com.example.medicare_call.service.CareCallSettingService;
+import com.example.medicare_call.service.carecall.CareCallSettingService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
