+++ src/main/java/com/example/medicare_call/controller/action/ElderHealthInfoController.java
@@ -4,8 +4,8 @@ import com.example.medicare_call.dto.CareCallSettingRequest;
 import com.example.medicare_call.dto.ElderHealthRegisterRequest;
 import com.example.medicare_call.dto.ElderHealthResponse;
 import com.example.medicare_call.global.annotation.AuthUser;
-import com.example.medicare_call.service.CareCallSettingService;
 import com.example.medicare_call.service.ElderHealthInfoService;
+import com.example.medicare_call.service.ElderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -36,4 +36,4 @@ public class ElderHealthInfoController {
     public ResponseEntity<List<ElderHealthResponse>> getElderHealthInfo(@Parameter(hidden = true) @AuthUser Long memberId) {
         return ResponseEntity.ok(elderHealthInfoService.getElderHealth(memberId.intValue()));
     }
-} 
+}
