+++ src/main/java/com/example/medicare_call/controller/action/TestCareCallController.java
@@ -1,7 +1,7 @@
 package com.example.medicare_call.controller.action;
 
 import com.example.medicare_call.dto.CareCallTestRequest;
-import com.example.medicare_call.service.CareCallService;
+import com.example.medicare_call.service.carecall.CareCallService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
