+++ src/main/java/com/example/medicare_call/scheduler/CareCallScheduler.java
@@ -1,7 +1,10 @@
 package com.example.medicare_call.scheduler;
 
-import com.example.medicare_call.service.CareCallSchedulerService;
+import com.example.medicare_call.domain.CareCallSetting;
+import com.example.medicare_call.repository.CareCallSettingRepository;
+import com.example.medicare_call.service.carecall.CareCallSchedulerService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
