+++ src/main/java/com/example/medicare_call/service/carecall/CareCallSchedulerService.java
@@ -1,4 +1,4 @@
-package com.example.medicare_call.service;
+package com.example.medicare_call.service.carecall;
 
 import com.example.medicare_call.domain.CareCallSetting;
 import com.example.medicare_call.global.enums.CallType;
