+++ src/main/java/com/example/medicare_call/service/carecall/CareCallSettingService.java
@@ -1,4 +1,4 @@
-package com.example.medicare_call.service;
+package com.example.medicare_call.service.carecall;
 
 import com.example.medicare_call.domain.CareCallSetting;
 import com.example.medicare_call.domain.Elder;
