+++ src/test/java/com/example/medicare_call/Integration/CareCallIntegrationTest.java
@@ -8,8 +8,11 @@ import com.example.medicare_call.global.enums.ResidenceType;
 import com.example.medicare_call.repository.CareCallSettingRepository;
 import com.example.medicare_call.repository.ElderRepository;
 import com.example.medicare_call.repository.MemberRepository;
-import com.example.medicare_call.service.CareCallSchedulerService;
-import com.example.medicare_call.service.CareCallService;
+import com.example.medicare_call.scheduler.CareCallScheduler;
+import com.example.medicare_call.service.carecall.CareCallSchedulerService;
+import com.example.medicare_call.service.carecall.CareCallService;
+import com.example.medicare_call.service.ElderService;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
