+++ src/test/java/com/example/medicare_call/controller/action/CallDataControllerTest.java
@@ -6,12 +6,14 @@ import com.example.medicare_call.domain.Elder;
 import com.example.medicare_call.dto.CallDataRequest;
 import com.example.medicare_call.domain.Member;
 import com.example.medicare_call.global.jwt.JwtProvider;
+import com.example.medicare_call.global.jwt.JwtUserDetails;
 import com.example.medicare_call.repository.CareCallRecordRepository;
 import com.example.medicare_call.repository.CareCallSettingRepository;
 import com.example.medicare_call.repository.ElderRepository;
 import com.example.medicare_call.repository.MemberRepository;
-import com.example.medicare_call.service.CallDataService;
+import com.example.medicare_call.service.carecall.CallDataService;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
